requirements.txt
Flask
psutil
requests
python-dotenv
